Move ActorManagerTests
